requirements.txt
flask==3.0.0
flask-cors==4.0.0
replicate==1.0.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
